Fix bug where unassigning a document was too aggressive and deleted relationships where it shouldn't.

+++ src/ifcopenshell-python/ifcopenshell/api/document/unassign_document.py
@@ -58,4 +58,7 @@ class Usecase:
     def execute(self):
         for rel in self.settings["product"].HasAssociations:
             if rel.is_a("IfcRelAssociatesDocument") and rel.RelatingDocument == self.settings["document"]:
-                self.file.remove(rel)
+                if len(rel.RelatedObjects) == 1:
+                    self.file.remove(rel)
+                else:
+                    rel.RelatedObjects = [o for o in rel.RelatedObjects if o != self.settings["product"]]
